refactor: move more funcs to root/file.c
