+++ root/file.c
@@ -2,6 +2,65 @@
  * Licensed under the Apache License, Version 2.0 */
 
 #include "watchman.h"
+#ifdef __APPLE__
+# include <sys/attr.h>
+#endif
+
+bool did_file_change(struct watchman_stat *saved, struct watchman_stat *fresh) {
+  /* we have to compare this way because the stat structure
+   * may contain fields that vary and that don't impact our
+   * understanding of the file */
+
+#define FIELD_CHG(name) \
+  if (saved->name != fresh->name) { \
+    return true; \
+  }
+
+  // Can't compare with memcmp due to padding and garbage in the struct
+  // on OpenBSD, which has a 32-bit tv_sec + 64-bit tv_nsec
+#define TIMESPEC_FIELD_CHG(wat) { \
+  struct timespec a = saved->wat##time; \
+  struct timespec b = fresh->wat##time; \
+  if (a.tv_sec != b.tv_sec || a.tv_nsec != b.tv_nsec) { \
+    return true; \
+  } \
+}
+
+  FIELD_CHG(mode);
+
+  if (!S_ISDIR(saved->mode)) {
+    FIELD_CHG(size);
+    FIELD_CHG(nlink);
+  }
+  FIELD_CHG(dev);
+  FIELD_CHG(ino);
+  FIELD_CHG(uid);
+  FIELD_CHG(gid);
+  // Don't care about st_blocks
+  // Don't care about st_blksize
+  // Don't care about st_atimespec
+  TIMESPEC_FIELD_CHG(m);
+  TIMESPEC_FIELD_CHG(c);
+
+  return false;
+}
+
+void struct_stat_to_watchman_stat(const struct stat *st,
+                                  struct watchman_stat *target) {
+  target->size = (off_t)st->st_size;
+  target->mode = st->st_mode;
+  target->uid = st->st_uid;
+  target->gid = st->st_gid;
+  target->ino = st->st_ino;
+  target->dev = st->st_dev;
+  target->nlink = st->st_nlink;
+  memcpy(&target->atime, &st->WATCHMAN_ST_TIMESPEC(a),
+      sizeof(target->atime));
+  memcpy(&target->mtime, &st->WATCHMAN_ST_TIMESPEC(m),
+      sizeof(target->mtime));
+  memcpy(&target->ctime, &st->WATCHMAN_ST_TIMESPEC(c),
+      sizeof(target->ctime));
+}
 
 static void watch_file(struct write_locked_watchman_root *lock,
                        struct watchman_file *file) {
